+++ backend/src/controllers/report.controller.js
@@ -34,21 +34,22 @@ const createReport = async (req, res) => {
 
 const resolveReport = async (req, res) => {
   try {
-    const userId = req.user
     const { id } = req.params
     const { status, resolution } = req.body
 
-    const user = await User.findById(userId)
+    const user = await User.findById(req.user)
     if (user == null || !user.isSuperAdmin) {
-      return res.status(403).json({ message: 'Unauthorized' })
+      res.status(403).json({ message: 'Unauthorized' })
+      return
     }
 
-    const report = await reportService.resolveReport(
-      user,
-      id,
-      status,
-      resolution
-    )
+    const report = await Report.findById(id)
+    if (report == null) {
+      res.status(400).json({ message: `Report with id ${id} not found.` })
+      return
+    }
+
+    await reportService.resolveReport(user, report, status, resolution)
 
     if (report == null)
       return res
@@ -89,26 +90,14 @@ const getReport = async (req, res) => {
     const { id } = req.params
 
     const user = await User.findById(userId)
-    if (user == null) {
-      return res.status(403).json({ message: 'Unauthorized' })
+    const report = (await reportService.getReport(id)).toJSON()
+
+    if (report == null || (report.user._id != userId && !user.isSuperAdmin)) {
+      res.status(400).json({ message: `Report with id ${id} not found.` })
+      return
     }
 
-    const report = (await reportService.getReport(id)).toJSON()
-    if (report == null || (report.user._id != userId && !user.isSuperAdmin))
-      res.status(400).json({ message: `Report with id ${id} not found.` })
-
     res.status(200).json(report)
-  } catch (error) {
-    res.status(500).json(error)
-    console.error(error)
-  }
-}
-
-const getUserReports = async (req, res) => {
-  try {
-    const user = req.user
-    const reports = await reportService.getUserReports(user)
-    res.status(200).json(reports)
   } catch (error) {
     res.status(500).json(error)
     console.error(error)
@@ -118,6 +107,13 @@ const getUserReports = async (req, res) => {
 const getEntityReports = async (req, res) => {
   try {
     const { entity_type, entity_id } = req.params
+
+    const user = await User.findById(req.user)
+    if (!user.isSuperAdmin) {
+      res.status(403).json({ message: 'Unauthorized' })
+      return
+    }
+
     const reports = await reportService.getEntityReports(entity_type, entity_id)
     res.status(200).json(reports)
   } catch (error) {
@@ -131,6 +127,5 @@ module.exports = {
   resolveReport,
   getAllReports,
   getReport,
-  getUserReports,
   getEntityReports
 }